git-svn-id: http://springhead.info/spr2/Springhead2/trunk@6436 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHFemMeshThermo.cpp
@@ -4409,7 +4409,7 @@ void PHFemMeshThermo::DecrMoist_velo(double vel){
 		double faceS=0.0;
 		for(int ii=0;ii<4;ii++){
 			int facet = tets[id].faces[ii];
-			if(facet <= nSurfaceFace){
+			if(facet <= (int)nSurfaceFace){
 				faceS += faces[facet].area;
 			}
 		}
@@ -4478,7 +4478,7 @@ void PHFemMeshThermo::DecrMoist_vel(double dt){
 		double faceS=0.0;
 		for(int ii=0;ii<4;ii++){
 			int facet = tets[id].faces[ii];
-			if(facet <= nSurfaceFace){
+			if(facet <= (int)nSurfaceFace){
 				faceS += faces[facet].area;
 			}
 		}
